klai-portal/backend/pyproject.toml
[project]
name = "klai-portal-api"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136",
    "uvicorn[standard]>=0.44",
    "httpx>=0.28",
    "pydantic[email]>=2.13",
    "pydantic-settings>=2.13",
    "asyncpg>=0.31",
    "sqlalchemy[asyncio]>=2.0.49",
    "alembic>=1.18",
    "python-multipart>=0.0.26",
    "docker>=7.1",
    "cryptography>=46.0",
    "icalendar>=6.1.0,<8.0",
    "motor>=3.7",
    "prometheus-client>=0.25,<1.0",
    "pyyaml>=6.0.3",
    "structlog>=25.5",
    "redis[hiredis]>=7.4",
    "pyjwt>=2.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=9",
    "pytest-asyncio>=1",
    "httpx>=0.28",
    "ruff>=0.15",
    "pyright>=1.1.408",
]

[dependency-groups]
dev = [
    "pytest>=9",
    "pytest-asyncio>=1",
    "httpx>=0.28",
    "ruff>=0.15",
    "pyright>=1.1.408",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [".", "tests"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "C90", "B", "TRY", "S", "RUF", "G", "LOG"]
ignore = [
    "E501",    # line length — handled by formatter
    "B008",    # Depends() in default args — standard FastAPI pattern, not a bug
    "S101",    # assert is fine in tests
    "S105",    # hardcoded passwords — too many false positives on config defaults
    "TRY003",  # long exception messages — perfectly fine
    "TRY300",  # return in try block — common in FastAPI
    "TRY301",  # raise in try block — common in FastAPI
    "TRY401",  # {e} in logger.exception() call — we want that context
]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.per-file-ignores]
"alembic/*" = ["UP", "I", "E501"]  # auto-generated migrations, don't enforce style
"tests/*" = ["S106"]  # hardcoded "passwords" in test fixtures (e.g. access_token="at-1") are fake tokens
